Rename utils.py to structured_logger.py
